# AV2/bazar-tem-tudo/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs do BazarTemTudo",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/estoques": {
            "get": {
                "tags": [
                    "Estoques"
                ],
                "summary": "Lista todos os estoques",
                "operationId": "ac5cfb4bb8f7a12c6f6459bd65764293",
                "responses": {
                    "200": {
                        "description": "Lista de estoques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "quantidade": {
                                                "type": "integer"
                                            },
                                            "id_produto": {
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "type": "string"
                                            },
                                            "updated_at": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Estoques"
                ],
                "summary": "Adicionar um novo item ao estoque",
                "operationId": "3ecfb48be7cfdb7f095095ac2e520653",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantidade": {
                                        "type": "integer"
                                    },
                                    "id_produto": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item adicionado ao estoque com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "quantidade": {
                                            "type": "integer"
                                        },
                                        "id_produto": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/estoques/{id}": {
            "get": {
                "tags": [
                    "Estoques"
                ],
                "summary": "Buscar item do estoque pelo ID",
                "operationId": "5f6b53324837052f1417b66ed2a29061",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do item do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item do estoque encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "quantidade": {
                                            "type": "integer"
                                        },
                                        "id_produto": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item do estoque não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Estoques"
                ],
                "summary": "Atualizar um item do estoque pelo ID",
                "operationId": "78703aacf8002fc34d59b5af482f8368",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do item do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "quantidade": {
                                        "type": "integer"
                                    },
                                    "id_produto": {
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "type": "string"
                                    },
                                    "updated_at": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item do estoque atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "quantidade": {
                                            "type": "integer"
                                        },
                                        "id_produto": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item do estoque não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estoques"
                ],
                "summary": "Remover um item do estoque pelo ID",
                "operationId": "49ebed0b5543a2840555aa6f2331662c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do item do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item do estoque removido com sucesso"
                    },
                    "404": {
                        "description": "Item do estoque não encontrado"
                    }
                }
            }
        },
...
    },
...
}